ViewHandle: make invalidation opt-in
When implementing the Surface API for native I made fuse forward InvalidateVisual to its native views. This has had some unexpected sideeffects, like Images with Mask suddenly disappearing. In that specific case Invalidate was called on the GraphicsView, which was not needed and probably messes up some internal GL state.

Making Invalidation opt-in, this restores the old behavior and adds the appropriate facilities for implementing the Surface API